refactor(noise-grid): extract shared GPU noise-field engine
Pull the reusable parts of the v11 GPU port into utils/noiseFieldGpu.js so
the rest of the noise-grid family can be converted without duplicating the
heavy plumbing:

- bit-exact p5 Perlin noise on the GPU (2D and 3D) via the packed
  permutation-table texture,
- a small GLSL stdlib (perlinNoise, rotateAroundCenter, paletteRainbow),
- the off-screen WebGL buffer, full-screen quad, and draw/composite cycle,
- a value-typed uniform setter and seed-managed noise texture.

A sketch now supplies only its fragment source (extra uniforms + main) and
drives it through render({ seed, octaves, falloff, columns, rows, center,
uniforms }). v11 is rewired onto it with identical shader logic, so its
output is unchanged; it drops from ~700 to ~120 lines. options.ts untouched.
